requirements.txt
feedparser>=6.0.0
requests>=2.28.0
pandas>=2.0.0
tqdm>=4.65.0
python-dateutil>=2.8.0
openai-whisper>=20231117
